LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/mismatched-snapshot-containing-block-size-skips-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Skipping view transition because viewport size changed.
CONSOLE MESSAGE: TypeError: e.data.startsWith is not a function. (In 'e.data.startsWith('FAIL:')', 'e.data.startsWith' is undefined)

PASS
View transitions: mismatched snapshot containing block size skips transition.
